% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR.R
\name{DR}
\alias{DR}
\title{Dose-response model function for listeriosis}
\usage{
DR(
  Dose,
  model = "JEMRA",
  population = 1,
  Poisson = FALSE,
  method = "integrate",
  ...
)
}
\arguments{
\item{Dose}{average dose in \code{CFU/serving} assuming a Poisson distribution from serving to serving (scalar or vector)
or exact dose if \code{Poisson} is \code{TRUE}}

\item{model}{either \code{JEMRA}, \code{Pouillot}, \code{Fritsch}, \code{EFSA}, \code{EFSAMV}, \code{EFSAV} or \code{EFSALV}}

\item{population}{considered population (scalar or vector) (see below)}

\item{Poisson}{if \code{TRUE}, assume that \code{Dose} is the mean of a Poisson distribution.
(actual LogNormal Poisson). If \code{FALSE} (default), assume that \code{Dose} is the actual number of bacteria.}

\item{method}{either \code{integrate} or \code{cubature} to specify the integration method.}

\item{...}{further arguments to pass to the \code{\link[=DRLogNormPoisson]{DRLogNormPoisson()}} function}
}
\value{
A vector of size Dose (if \code{population} is a scalar) or a matrix of
dimension (length of the \code{Dose} vector times length of the \code{population} vector)
}
\description{
The \code{\link[=DR]{DR()}} function provides the marginal probability of invasive listeriosis
in a given \code{population} for a given \code{Dose} in CFU using the
\code{JEMRA}, the \code{Pouillot}, the \code{Fritsch}, the \code{EFSA} dose-response models
or the model developed within the "WHO Risk assessment of \emph{Listeria monocytogenes} in foods" project (\code{EFSAMV}, \code{EFSAV}, \code{EFSALV}) (see References).
}
\details{
For the \code{Pouillot}, \code{Fritsch}, \code{EFSA} \code{EFSAMV},\code{EFSAV} or \code{EFSALV} model,
it integrates a log10Normal Poisson distribution (\insertCite{Pouillot2015;textual}{doseresponsemodels})
or a log10Normal-binomial distribution. For the \code{JEMRA} model
an exponential dose-response model or a binomial model is used.

The marginal dose-Response over all sub-populations (population: 0) is obtained by weighting
the estimated probabilities according to the proportion of each sub-populations
as provided by \insertCite{FAO-WHO2004}{doseresponsemodels} (82.5\% healthy, 17.5\%
with increased susceptibility), \insertCite{Pouillot2015;textual}{doseresponsemodels},
(their Table I, based on frequency of each sub-populations in France),
\insertCite{Fritsch2018}{doseresponsemodels}, and \insertCite{EFSA2018}{doseresponsemodels}
(their Table 2, based on the number of eating occasions per sub-population).

This function is slow. Use \code{\link[=DRQuick]{DRQuick()}} in production.\tabular{lll}{
   Model \tab Population \tab Characteristics \cr
   JEMRA \tab 0 \tab Marginal over subpopulations \cr
   JEMRA \tab 1 \tab Healthy population \cr
   JEMRA \tab 2 \tab Increased susceptibility \cr
   Pouillot \tab 0 \tab Marginal over subpopulations \cr
   Pouillot \tab 1 \tab Less than 65 years old \cr
   Pouillot \tab 2 \tab More than 65 years old \cr
   Pouillot \tab 3 \tab Pregnancy \cr
   Pouillot \tab 4 \tab Nonhematological Cancer \cr
   Pouillot \tab 5 \tab Hematological cancer \cr
   Pouillot \tab 6 \tab Renal or Liver failure \cr
   Pouillot \tab 7 \tab Solid organ transplant \cr
   Pouillot \tab 8 \tab Inflammatory diseases \cr
   Pouillot \tab 9 \tab HIV/AIDS \cr
   Pouillot \tab 10 \tab Diabetes \cr
   Pouillot \tab 11 \tab Hear diseases \cr
   Fritsch \tab 0 \tab Marginal over virulence \cr
   Fritsch \tab 1 \tab Highly virulent \cr
   Fritsch \tab 2 \tab Medium virulent \cr
   Fritsch \tab 3 \tab Hypovirulent \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 0 \tab Marginal over subpopulations \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 1 \tab Female 1-4 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 2 \tab Male 1-4 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 3 \tab Female 5-14 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 4 \tab Male 5-14 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 5 \tab Female 15-24 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 6 \tab Male 15-24 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 7 \tab Female 25-44 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 8 \tab Male 25-44 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 9 \tab Female 45-64 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 10 \tab Male 45-64 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 11 \tab Female 65-74 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 12 \tab Male 65-74 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 13 \tab Female >75 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 14 \tab Male >75 yo \cr
}


See the parameters in the JEMRA report.

\code{method = "cubature"} will use the \link[cubature]{hcubature} function that
is much slower but guarantees a tolerance of \eqn{1E-5}.
}
\examples{
DR(5:10, "Pouillot", 1:11)
DR(5:10, "Pouillot", 0)

}
\references{
\insertRef{EFSA2018}{doseresponsemodels}

\insertRef{FAO-WHO2004}{doseresponsemodels}

\insertRef{Fritsch2018}{doseresponsemodels}

c
}
\author{
Regis Pouillot
}
\keyword{dose-response}
\keyword{model}
