% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRLogNormPoisson.R
\name{DRLogNormPoisson}
\alias{DRLogNormPoisson}
\title{Marginal probability of invasive listeriosis over r}
\usage{
DRLogNormPoisson(
  Dose,
  meanlog10,
  sdlog10,
  Poisson = FALSE,
  low = -Inf,
  up = Inf,
  silent = TRUE,
  tol = 1e-20,
  method = "integrate",
  ...
)
}
\arguments{
\item{Dose}{(\code{CFU/serving}) Dose (scalar or vector). It should be integers if \code{Poisson} is \code{TRUE}.}

\item{meanlog10}{the meanlog10 parameter of the distribution of \code{r} (parameter of the exponential model).}

\item{sdlog10}{the sdlog10 parameter of the distribution of \code{r} (parameter of the exponential model).}

\item{Poisson}{if \code{TRUE}, assume that \code{Dose} is the mean of a Poisson distribution. (actual LogNormal Poisson).
If \code{FALSE} (default), assume that \code{Dose} is the actual number of bacteria.}

\item{low}{lower value for the integration.}

\item{up}{upper value for the integration.}

\item{silent}{silent the error-try function.}

\item{tol}{relative tolerance. Note: for \code{method = "cubature"}, the tolerance will be set to \eqn{1E-05}.}

\item{method}{either \code{"integrate"} (default) or \code{"cubature"} to specify the integration method.}

\item{...}{further arguments to pass to the integrate function.}
}
\value{
Probability of invasive listeriosis integrated over \code{r}.
}
\description{
The function \code{\link[=DRLogNormPoisson]{DRLogNormPoisson()}} provides the marginal probability of invasive listeriosis
in a given population for a given \code{Dose} in \code{CFU}. this function is not vectorized.
}
\details{
The function evaluates
\deqn{\int_{low}^{inf} \Phi(x, mulog_{10}, sdlog_{10})\cdot(1-e^{(-Dose \cdot 10^{r})}) dr}
using the \code{\link{integrate}} function, with a relative tolerance equals to \code{tol}
if \code{Poisson} is \code{TRUE}. If \code{Poisson} is \code{FALSE}, it evaluates
\deqn{\int_{low}^{inf} \Phi(x, mulog_{10}, sdlog_{10})\cdot(1-(1-10^{r})^{Dose}) dr}.

For \code{method = "cubature"}, the tolerance will be set to \eqn{1E-5}.
\code{method = "cubature"} will use the \verb{\link[cubature]\{hcubature\}} function that
is much slower but guarantees a tolerance of \eqn{1E-5}.
}
\note{
This function is used by the \code{\link[=DR]{DR()}} function, a wrapper of \code{\link[=DRLogNormPoisson]{DRLogNormPoisson()}}.
For a quick, vectorized version of it, use \code{\link[=DRQuick]{DRQuick()}}.
}
\examples{
DRLogNormPoisson(2, -14.11, 1.62,low=-Inf,up=Inf)
DRLogNormPoisson(2, -14.11, 1.62, Poisson=TRUE)

}
\references{
\insertRef{Pouillot2015}{doseresponsemodels}
}
\author{
Régis Pouillot
}
\keyword{dose-response}
\keyword{model}
