% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRQuick.R
\name{DRQuick}
\alias{DRQuick}
\title{Dose-response model function for listeriosis - quick version}
\usage{
DRQuick(Dose, model = "JEMRA", population = 1, Poisson = FALSE)
}
\arguments{
\item{Dose}{(\code{CFU/serving}) Dose (scalar or vector).}

\item{model}{either \code{JEMRA}, \code{Pouillot}, \code{Fritsch}, \code{EFSA}, \code{EFSAMV},\code{EFSAV} or \code{EFSALV}}

\item{population}{considered population (scalar or vector).}

\item{Poisson}{if \code{TRUE}, assume that \code{Dose} is the mean of a Poisson distribution.
(actual LogNormal Poisson). If \code{FALSE} (default), assume that \code{Dose} is the actual number of bacteria.}
}
\value{
A vector of size \code{Dose} (if \code{population} is a scalar) or a matrix of
dimension (length of the \code{Dose} vector x length of the \code{population} vector)
}
\description{
This function provides the marginal probability of invasive listeriosis
in a given \code{population} for a given \code{Dose} in \code{CFU} using the
\code{JEMRA}, the \code{Pouillot}, the \code{Fritsch} or the \code{EFSA} dose-response models
or the model developed within this project (\code{EFSAMV},\code{EFSAV},\code{EFSALV}) (see References).
}
\details{
\tabular{lll}{
   Model \tab Population \tab Characteristics \cr
   JEMRA \tab 1 \tab Healthy population \cr
   JEMRA \tab 2 \tab Increased susceptibility \cr
   Pouillot \tab 1 \tab Less than 65 years old \cr
   Pouillot \tab 2 \tab More than 65 years old \cr
   Pouillot \tab 3 \tab Pregnancy \cr
   Pouillot \tab 4 \tab Nonhematological Cancer \cr
   Pouillot \tab 5 \tab Hematological cancer \cr
   Pouillot \tab 6 \tab Renal or Liver failure \cr
   Pouillot \tab 7 \tab Solid organ transplant \cr
   Pouillot \tab 8 \tab Inflammatory diseases \cr
   Pouillot \tab 9 \tab HIV/AIDS \cr
   Pouillot \tab 10 \tab Diabetes \cr
   Pouillot \tab 11 \tab Hear diseases \cr
   Fritsch \tab 1 \tab Highly virulent \cr
   Fritsch \tab 2 \tab Medium virulent \cr
   Fritsch \tab 3 \tab Hypovirulent \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 1 \tab Female 1-4 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 2 \tab Male 1-4 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 3 \tab Female 5-14 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 4 \tab Male 5-14 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 5 \tab Female 15-24 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 6 \tab Male 15-24 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 7 \tab Female 25-44 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 8 \tab Male 25-44 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 9 \tab Female 45-64 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 10 \tab Male 45-64 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 11 \tab Female 65-74 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 12 \tab Male 65-74 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 13 \tab Female >75 yo \cr
   EFSA-EFSALV-EFSAV-EFSAMV \tab 14 \tab Male >75 yo \cr
}


See the parameters in the JEMRA report.
}
\note{
This function uses (for all model but \code{JEMRA}) a linear approximation (\code{approxfun})
from the exact \code{\link[=DR]{DR()}} model evaluated on \eqn{Dose = c(0,10^{seq(-5,12,length=1701)})}
(if \code{Poisson=TRUE}) or \eqn{c(0,10^{seq(0,12,length=2000)})} (if \code{Poisson=FALSE}).
Any Dose lower or higher than these ranges will lead to \code{NA}.
}
\examples{
# Compare DR and DRQuick
cbind(DR(1:10, model="Pouillot", population = 5), 
      DRQuick(1:10,  model="Pouillot", population = 5))
DRQuick(1:10,  model="Pouillot", population = 2)

}
\references{
\insertRef{EFSA2018}{doseresponsemodels}

\insertRef{FAO-WHO2004}{doseresponsemodels}

\insertRef{Fritsch2018}{doseresponsemodels}

\insertRef{Pouillot2015}{doseresponsemodels}
}
\seealso{
\code{\link[=DR]{DR()}},  \code{\link[=DRLogNormPoisson]{DRLogNormPoisson()}}.
}
\author{
Regis Pouillot
}
\keyword{dose-response}
\keyword{model}
