% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRLm.R
\name{doseresponsemodels-package}
\alias{doseresponsemodels-package}
\alias{doseresponsemodels}
\alias{#@references}
\title{doseresponsemodels}
\description{
The doseresponsemodels package provides Dose Response functions for Listeria monocytogenes.
}
\section{doseresponsemodels functions}{

The doseresponsemodels functions ...
}

\keyword{#@examples}
\keyword{package}
