% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseresponsemodels-package.R
\docType{package}
\name{doseresponsemodels-package}
\alias{doseresponsemodels}
\alias{doseresponsemodels-package}
\title{doseresponsemodels: Dose-Response models currently for Listeria monocytogenes}
\description{
This package provides the functions of the Dose-Response model for Listeria monocytogenes from WHO 2024.
}
\author{
\strong{Maintainer}: Régis Pouillot \email{rpouillot.work@gmail.com} (\href{https://orcid.org/0000-0002-6107-5212}{ORCID})

Authors:
\itemize{
  \item Andrea Kiermeier \email{andreas.kiermeier@gmail.com}
  \item Vasco Cadavez \email{vcadavez@ipb.pt} (\href{https://orcid.org/0000-0002-3077-7414}{ORCID})
  \item Ursula Gonzales-Barron \email{ubarron@ipb.pt} (\href{https://orcid.org/0000-0002-8462-9775}{ORCID})
  \item Laurent Guillier \email{Laurent.guillier@anses.fr} (\href{https://orcid.org/0000-0002-7867-2937}{ORCID})
  \item Moez Sanaa \email{sanaam@who.int} (\href{https://orcid.org/0000-0003-3576-0379}{ORCID})
}

}
\keyword{internal}
